+++ src/tests/bucketlistItems.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import BucketlistItems from "../components/Bucketlist/bucketlistItems";
-import BLIRow from "../components/Bucketlist/bucketlistItems";
+import BLIRow from "../components/Bucketlist/BLIRow";
 
 describe('bucketlist items snapshot', () => {
     it('matches snapshot', () => {
